Separate R2 registration form handling from Finna forms.
