+++ local/config/finna/FeedbackForms.yaml.sample
@@ -214,121 +214,3 @@ forms:
         type: textarea
         label: feedback_message
         required: true
-
-  # REMS registration form for new users
-  R2Register:
-    allowLocalOverride: false
-    title: "R2_register_form_title"
-    enabled: false
-    useCaptcha: false
-    onlyForLoggedUsers: true
-    response: Thank you for your feedback.
-    senderInfoRequired: false
-    senderInfoHelp: feedback_info
-    hideSenderInfo: true
-    hideRecipientInfo: true
-    help:
-      pre: R2_register_form_help_pre_html
-      post: R2_register_form_help_post_html
-
-    fields:
-      - name: usage_purpose
-        type: select
-        options:
-          - R2_register_form_usage_1
-          - R2_register_form_usage_2
-          - R2_register_form_usage_3
-          - R2_register_form_usage_4
-          - R2_register_form_usage_5
-          - R2_register_form_usage_6
-        label: R2_register_form_usage
-        help:
-          post: R2_register_form_usage_help_html
-
-        placeholder: Select
-        required: true
-      - name: firstname
-        type: text
-        required: true
-        label: First Name
-        help: R2_register_form_userinfo
-
-      - name: lastname
-        type: text
-        required: true
-        label: Last Name
-      - name: email
-        type: email
-        required: true
-        label: Email
-        help:
-          post: R2_register_form_datasource
-      - name: license
-        type: checkbox
-        options:
-          - R2_register_form_license_accept
-        required: true
-      - name: age
-        type: checkbox
-        options:
-          - R2_register_form_age
-        required: true
-
-  # REMS registration form for returning users
-  R2RegisterReturningUser:
-    allowLocalOverride: false
-    title: "R2_register_form_title"
-    enabled: false
-    useCaptcha: false
-    onlyForLoggedUsers: true
-    response: Thank you for your feedback.
-    senderInfoRequired: false
-    senderInfoHelp: feedback_info
-    hideSenderInfo: true
-    hideRecipientInfo: true
-    help:
-      pre: R2_register_form_help_returninguser_pre_html
-      post: R2_register_form_help_post_html
-
-    fields:
-      - name: usage_purpose
-        type: select
-        options:
-          - R2_register_form_usage_1
-          - R2_register_form_usage_2
-          - R2_register_form_usage_3
-          - R2_register_form_usage_4
-          - R2_register_form_usage_5
-          - R2_register_form_usage_6
-        label: R2_register_form_usage
-        help:
-          post: R2_register_form_usage_help_html
-
-        placeholder: Select
-        required: true
-      - name: firstname
-        type: text
-        required: true
-        label: First Name
-        help: R2_register_form_userinfo
-
-      - name: lastname
-        type: text
-        required: true
-        label: Last Name
-      - name: email
-        type: email
-        required: true
-        label: Email
-        help:
-          post: R2_register_form_datasource
-      - name: license
-        type: checkbox
-        options:
-          - R2_register_form_license_accept
-        required: true
-      - name: age
-        type: checkbox
-        options:
-          - R2_register_form_age
-        required: true
