+++ local/config/finna/R2FeedbackForms.yaml.sample
@@ -1,0 +1,180 @@
+---
+# Listing of configured forms.
+#
+# Configured forms are listed in the top level 'forms' array and can accessed via the URL:
+# /Feedback/Form/<form-id>
+#
+# The main Feedback form (id: FeedbackSite) can be accessed with a shorter URL /Feedback
+#
+# Form supports the following options:
+#
+#   enabled (boolean) Is the form enabled? (default = true)
+#   useCaptcha (boolean) Should the form use Captcha validation? Requires the "feedback"
+#                        option to be turned on in the config.ini [Captcha] form setting.
+#                        (default = true, if Captcha turned on for feedback overall).
+#   title (string) Form title (translation key)
+#   onlyForLoggedUsers (boolean) Require the user to be logged in to see the form
+#                                (default = false)
+#   emailSubject (string) Email subject (translation key). (default = email_subject
+#                         value from config.ini [Feedback] section, or 'VuFind Feedback')
+#                         The subject may contain placeholders (for example %%receiver%%)
+#                         that are replaced with submitted form values.
+#   recipient
+#     name (string)  Recipient name (default = recipient_name setting from config.ini
+#                    [Feedback] section)
+#     email (string) Recipient email address (default = recipient_email setting from
+#                    config.ini [Feedback] section)
+#   response (string) Reponse after form submit (translation key)
+#   senderInfoRequired (boolean)  Require sender to fill out name and email fields
+#   senderInfoHelp (string) Info text for sender name/email (translation key)
+#
+#   help
+#     pre (string)  Help text displayed at the beginning of the form (translation key).
+#     post (string) Help text displayed at the end of the form (translation key)
+#
+#   Help texts may contain HTML-formatting.
+#
+#   fields (array) List of form elements with the following options:
+#
+#     name (string)        Element name
+#     label (string)       Element label (translation key)
+#     required (boolean)   Is the element required?
+#     settings (array)     HTML attributes as key-value pairs, for example:
+#       - [class, "custom-css-class another-class"]
+#     type (string)        Element type (text|textarea|email|url|select)
+#     help (string)        Element help text (translation key)
+#     helpExtended (array) 
+#       pre (string)       Element help text (translation key)
+#       post (string)
+#     
+#     And for select elements one of:
+# 
+#     options (array) List of select values (translation keys)
+#
+#     or
+#
+#     optionGroups (array) List of option groups with keys:
+#       label (string) Group label (translation key)
+#       options (array) List of select values (translation keys)
+#
+#-----------------------------------------------------------------------------------
+
+forms:
+
+   # REMS registration form for new users
+  R2Register:
+    allowLocalOverride: false
+    title: "R2_register_form_title"
+    enabled: false
+    useCaptcha: false
+    onlyForLoggedUsers: true
+    response: Thank you for your feedback.
+    senderInfoRequired: false
+    senderInfoHelp: feedback_info
+    hideSenderInfo: true
+    hideRecipientInfo: true
+    help:
+      pre: R2_register_form_help_pre_html
+      post: R2_register_form_help_post_html
+
+    fields:
+      - name: usage_purpose
+        type: select
+        options:
+          - R2_register_form_usage_1
+          - R2_register_form_usage_2
+          - R2_register_form_usage_3
+          - R2_register_form_usage_4
+          - R2_register_form_usage_5
+          - R2_register_form_usage_6
+        label: R2_register_form_usage
+        help:
+          post: R2_register_form_usage_help_html
+
+        placeholder: Select
+        required: true
+      - name: firstname
+        type: text
+        required: true
+        label: First Name
+        help: R2_register_form_userinfo
+
+      - name: lastname
+        type: text
+        required: true
+        label: Last Name
+      - name: email
+        type: email
+        required: true
+        label: Email
+        help:
+          post: R2_register_form_datasource
+      - name: license
+        type: checkbox
+        options:
+          - R2_register_form_license_accept
+        required: true
+      - name: age
+        type: checkbox
+        options:
+          - R2_register_form_age
+        required: true
+
+  # REMS registration form for returning users
+  R2RegisterReturningUser:
+    allowLocalOverride: false
+    title: "R2_register_form_title"
+    enabled: false
+    useCaptcha: false
+    onlyForLoggedUsers: true
+    response: Thank you for your feedback.
+    senderInfoRequired: false
+    senderInfoHelp: feedback_info
+    hideSenderInfo: true
+    hideRecipientInfo: true
+    help:
+      pre: R2_register_form_help_returninguser_pre_html
+      post: R2_register_form_help_post_html
+
+    fields:
+      - name: usage_purpose
+        type: select
+        options:
+          - R2_register_form_usage_1
+          - R2_register_form_usage_2
+          - R2_register_form_usage_3
+          - R2_register_form_usage_4
+          - R2_register_form_usage_5
+          - R2_register_form_usage_6
+        label: R2_register_form_usage
+        help:
+          post: R2_register_form_usage_help_html
+
+        placeholder: Select
+        required: true
+      - name: firstname
+        type: text
+        required: true
+        label: First Name
+        help: R2_register_form_userinfo
+
+      - name: lastname
+        type: text
+        required: true
+        label: Last Name
+      - name: email
+        type: email
+        required: true
+        label: Email
+        help:
+          post: R2_register_form_datasource
+      - name: license
+        type: checkbox
+        options:
+          - R2_register_form_license_accept
+        required: true
+      - name: age
+        type: checkbox
+        options:
+          - R2_register_form_age
+        required: true
