+++ module/Finna/src/Finna/Controller/R2ControllerTrait.php
@@ -40,19 +40,6 @@ use Laminas\ServiceManager\ServiceLocatorInterface;
  */
 trait R2ControllerTrait
 {
-    use \VuFind\Log\LoggerAwareTrait;
-
-    /**
-     * Constructor
-     *
-     * @param ServiceLocatorInterface $sm Service locator
-     */
-    public function __construct(ServiceLocatorInterface $sm)
-    {
-        $this->setLogger($sm->get('VuFind\Logger'));
-        parent::__construct($sm);
-    }
-
     /**
      * Handle onDispatch event
      *
@@ -68,162 +55,5 @@ trait R2ControllerTrait
         }
 
         return parent::onDispatch($e);
-    }
-
-    /**
-     * Replace R2 new user registration form id with the id for returning
-     * user registration form.
-     *
-     * @param string $formId Current form id
-     *
-     * @return null|string registration form id for returning user or null
-     * if the form id was not modified.
-     */
-    protected function replaceR2RegisterFormId($formId)
-    {
-        if (!\Finna\Form\Form::isR2RegisterForm($formId, true)) {
-            // Not a R2 registration form for new users
-            return null;
-        }
-
-        // R2 registration form requested.
-        // 1. If the user is not logged, default to new user form
-        //    (this method gets called again after the login).
-        // 2. For logged users, check if the user has been registered to REMS
-        //    and replace form id with the id for returning users.
-        if (!($user = $this->getUser())) {
-            return $formId;
-        }
-
-        $rems = $this->serviceLocator->get('Finna\Service\RemsService');
-        try {
-            $regId
-                = \Finna\Form\Form::getR2RegisterFormId(!$rems->isUserRegistered());
-        } catch (\Exception $e) {
-            return null;
-        }
-        return $formId !== $regId ? $regId : null;
-    }
-
-    /**
-     * Handles display and submit of R2 registration form.
-     *
-     * @return void
-     */
-    protected function processR2RegisterForm()
-    {
-        $formId = $this->params()->fromRoute('id', $this->params()->fromQuery('id'));
-        if (!\Finna\Form\Form::isR2RegisterForm($formId)) {
-            return null;
-        }
-
-        $inLightbox
-            = $this->getRequest()->getQuery('layout', 'no') === 'lightbox'
-               || 'layout/lightbox' == $this->layout()->getTemplate();
-
-        $getRedirect = function () use ($inLightbox) {
-            // Logged but not authorized (wrong login method etc), close form
-            if ($inLightbox) {
-                // Login completed inside lightbox: refresh page
-                $response = $this->getResponse();
-                $response->setStatusCode(205);
-                return '';
-            } else {
-                return $this->redirect()->toRoute('search-home');
-            }
-        };
-
-        if (!($user = $this->getUser())) {
-            // Not logged, prompt login
-            return $this->forceLogin();
-        }
-
-        // Verify that user is authenticated to access restricted R2 data.
-        if (!$this->isAuthenticated()) {
-            return $getRedirect();
-        }
-
-        // Check user permission from REMS and show registration if needed.
-        $rems = $this->serviceLocator->get('Finna\Service\RemsService');
-        try {
-            if ($rems->isUserBlocklisted()) {
-                return $getRedirect();
-            }
-        } catch (\Exception $e) {
-            $this->flashMessenger()->addErrorMessage('R2_rems_connect_error');
-            return $getRedirect();
-        }
-
-        try {
-            if ($rems->hasUserAccess(true)) {
-                // User already has access
-                return $getRedirect();
-            }
-        } catch (\Exception $e) {
-            $this->flashMessenger()->addErrorMessage('R2_rems_connect_error');
-            return $getRedirect();
-        }
-
-        if ($this->formWasSubmitted('submit')) {
-            // Handle submitted registration form
-            $form = $this->serviceLocator->get('VuFind\Form\Form');
-            $form->setFormId($formId);
-
-            $view = $this->createViewModel(compact('form', 'formId', 'user'));
-            $view->setTemplate('feedback/form');
-            $params = $this->params()->fromPost();
-            $form->setData($params);
-
-            if (!$form->isValid()) {
-                return $view;
-            }
-
-            $rems = $this->serviceLocator->get('Finna\Service\RemsService');
-
-            // Collect submitted params required by REMS form
-            $formParams = [];
-            $formParams['usage_purpose']
-                = (string)substr($params['usage_purpose'], -1);
-            $formParams['usage_purpose_text'] = $params['usage_purpose'];
-            if ($age = $params['age'] ?? null) {
-                $formParams['age'] = '1';
-            }
-            if ($license = $params['license'] ?? null) {
-                $formParams['license'] = '1';
-            }
-
-            // Take firstname and lastname from profile
-            $firstname = $user->firstname;
-            $lastname = $user->lastname;
-            $email = $params['email'];
-
-            try {
-                $rems->registerUser(
-                    $email,
-                    $firstname,
-                    $lastname,
-                    $formParams
-                );
-            } catch (\Exception $e) {
-                $this->flashMessenger()->addErrorMessage('R2_register_error');
-                $this->logError('REMS registration error: ' . $e->getMessage());
-            }
-
-            return $getRedirect();
-        }
-
-        // User is authorized, let parent display the registration form
-        return null;
-    }
-
-    /**
-     * Is the user authenticated to use R2?
-     *
-     * @return bool
-     */
-    protected function isAuthenticated()
-    {
-        return $this->serviceLocator->get(\Finna\Service\R2Service::class)
-            ->isAuthenticated();
     }
 }
